refactor: move rowsChanged into @shared/db-utils
The rows-affected fix landed as a private helper in `osn/api/src/lib/`, next
to three copies of the same function that `cire/api` had already grown — one
in `session.ts`, one in `retention.ts`, one declared inside an `Effect.gen`
body in `directory.ts`. Four copies of a driver-shape table is three too many:
the next driver added has to be found in four places, and the bug the helper
exists to prevent already shipped to production once.

`rowsChanged` now sits in `@shared/db-utils` beside `commitBatch` and
`dbQuery`, and all five call sites read it from there. Its unit tests move to
`shared/db-utils/tests/`; the osn regression tests that drive each CAS gate
through a D1-shaped driver proxy stay where the gates are.

`@cire/api` picks up a `workspace:*` dependency on `@shared/db-utils`.
No behaviour change.
