+++ cire/api/package.json
@@ -12,12 +12,13 @@
   },
   "dependencies": {
     "@cire/db": "workspace:*",
+    "@cire/invite-designs": "workspace:*",
     "@cire/theme": "workspace:*",
     "@elysiajs/cors": "^1.4.2",
     "@shared/crypto": "workspace:*",
+    "@shared/db-utils": "workspace:*",
     "@shared/email": "workspace:*",
     "@shared/feature-flags": "workspace:*",
-    "@cire/invite-designs": "workspace:*",
     "@shared/observability": "workspace:*",
     "@shared/osn-auth-client": "workspace:*",
     "@shared/rate-limit": "workspace:*",
