+++ wiki/changelog/security-fixes.md
@@ -16,7 +16,7 @@ Archived completed security findings from [[TODO]]. Finding IDs follow the [[rev
 
 ## Rows-affected read on D1 — every refresh killed its own session (2026-07-26)
 
-- **S-H (d1-rows-changed)** — **every production refresh grant deleted the session it was renewing and answered `400 invalid_grant`.** **Issue:** Drizzle reports rows affected differently per driver — bun:sqlite `{ changes }`, libsql `{ rowsAffected }`, Cloudflare D1 `{ success, meta: { changes } }`. Two compare-and-swap gates read `changes ?? rowsAffected ?? 0`. Tests run on bun:sqlite, production runs on D1, so both gates were green in CI and read 0 for every write in production. **Why:** in refresh rotation the old-session `DELETE` *is* the CAS, and 0 means "a concurrent grant won". Reading 0 every time meant the row was really gone, the replacement INSERT never ran, and the grant failed. Access tokens live five minutes, so every session died at its first refresh — this was the true root of the "logged out for no reason" reports, which earlier work ([[sessions]] grace window) had only narrowed. Production bore it out: no session row had ever had `last_used_at` move past `created_at`. Passkey **rename** carried the same read, so a rename that updated the row still answered "Passkey not found". **Solution:** one helper, `osn/api/src/lib/rows-changed.ts`, that knows all three shapes; both gates go through it, and any new rows-affected check must too. Regression tests drive each gate through a driver proxy that reports counts D1-style — the shape difference, not the logic, is what the old tests could not see. See [[sessions]].
+- **S-H (d1-rows-changed)** — **every production refresh grant deleted the session it was renewing and answered `400 invalid_grant`.** **Issue:** Drizzle reports rows affected differently per driver — bun:sqlite `{ changes }`, libsql `{ rowsAffected }`, Cloudflare D1 `{ success, meta: { changes } }`. Two compare-and-swap gates read `changes ?? rowsAffected ?? 0`. Tests run on bun:sqlite, production runs on D1, so both gates were green in CI and read 0 for every write in production. **Why:** in refresh rotation the old-session `DELETE` *is* the CAS, and 0 means "a concurrent grant won". Reading 0 every time meant the row was really gone, the replacement INSERT never ran, and the grant failed. Access tokens live five minutes, so every session died at its first refresh — this was the true root of the "logged out for no reason" reports, which earlier work ([[sessions]] grace window) had only narrowed. Production bore it out: no session row had ever had `last_used_at` move past `created_at`. Passkey **rename** carried the same read, so a rename that updated the row still answered "Passkey not found". **Solution:** one helper, `rowsChanged` in `@shared/db-utils`, that knows all three shapes; both gates go through it, and any new rows-affected check must too. `cire/api` had grown three private copies of the same function — they now read the shared one. Regression tests drive each gate through a driver proxy that reports counts D1-style — the shape difference, not the logic, is what the old tests could not see. See [[sessions]].
 
 ## Access-token session binding — `osn_sid` (2026-07-26)
